requirements.txt
Pillow==11.2.1numpy==2.2.5